# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "High-performance backend API for HiDocs Form \u0026 Exam Maker application using Golang and Domain-Driven Design (DDD).",
        "title": "HiDocs Backend API (Form \u0026 Exam Maker)",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "HiDocs Support",
            "email": "support@hidocs.id"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/admin/creators": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "List form creators/users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.CreatorResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Create a new form creator account",
                "parameters": [
                    {
                        "description": "Create Creator Payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.UserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/creators/{creator_id}/status": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Update creator active/inactive status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Creator ID",
                        "name": "creator_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status Payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateCreatorStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/dashboard/stats": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get admin dashboard global statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.AdminStats"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/forms": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "List all system forms",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.FormResponseDTO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/forms/{form_id}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete form as admin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Form ID",
                        "name": "form_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
        "domain.AdminStats": {
            "type": "object",
            "properties": {
                "active_exams": {
                    "type": "integer"
                },
                "total_creators": {
                    "type": "integer"
                },
                "total_forms": {
                    "type": "integer"
                },
                "total_responses": {
                    "type": "integer"
                },
                "total_users": {
                    "type": "integer"
                }
            }
        },
...
        "domain.FormSettings": {
            "type": "object",
            "properties": {
                "auto_active_days": {
                    "type": "integer"
                },
                "duration_minutes": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "form_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active_immediately": {
                    "type": "boolean"
                },
                "is_one_time_submission": {
                    "type": "boolean"
                },
                "randomize_options": {
                    "type": "boolean"
                },
                "randomize_questions": {
                    "type": "boolean"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "domain.FormStatus": {
            "type": "string",
            "enum": [
                "DRAFT",
                "ACTIVE",
                "CLOSED"
            ],
            "x-enum-varnames": [
                "StatusDraft",
                "StatusActive",
                "StatusClosed"
            ]
        },
        "domain.FormType": {
            "type": "string",
            "enum": [
                "SURVEY",
                "EXAM"
            ],
            "x-enum-varnames": [
                "TypeSurvey",
                "TypeExam"
            ]
        },
...
        "domain.QuestionType": {
            "type": "string",
            "enum": [
                "SHORT_TEXT",
                "LONG_TEXT",
                "MULTIPLE_CHOICE",
                "CHECKBOXES",
                "DROPDOWN",
                "RATING",
                "YES_NO",
                "MATH",
                "CODE",
                "IMAGE"
            ],
            "x-enum-varnames": [
                "TypeShortText",
                "TypeLongText",
                "TypeMultipleChoice",
                "TypeCheckboxes",
                "TypeDropdown",
                "TypeRating",
                "TypeYesNo",
                "TypeMath",
                "TypeCode",
                "TypeImage"
            ]
        },
        "domain.UserRole": {
            "type": "string",
            "enum": [
                "admin",
                "user"
            ],
            "x-enum-varnames": [
                "RoleAdmin",
                "RoleUser"
            ]
        },
...
        "dto.CreatorResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "dto.FormResponseDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "custom_url": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "form_settings": {
                    "$ref": "#/definitions/domain.FormSettings"
                },
                "id": {
                    "type": "string"
                },
                "is_template": {
                    "type": "boolean"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.QuestionDTO"
                    }
                },
                "response_count": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/domain.FormStatus"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/domain.FormType"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.OptionDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "is_correct": {
                    "type": "boolean"
                },
                "option_text": {
                    "type": "string"
                },
                "order_index": {
                    "type": "integer"
                },
                "question_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.QuestionDTO": {
            "type": "object",
            "properties": {
                "code_language": {
                    "type": "string"
                },
                "form_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "img_url": {
                    "type": "string"
                },
                "is_auto_scored": {
                    "type": "boolean"
                },
                "is_required": {
                    "type": "boolean"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.OptionDTO"
                    }
                },
                "order_index": {
                    "type": "integer"
                },
                "points": {
                    "type": "integer"
                },
                "question_text": {
                    "type": "string"
                },
                "question_type": {
                    "$ref": "#/definitions/domain.QuestionType"
                }
            }
        },
        "dto.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                }
            }
        },
...
        "dto.UpdateCreatorStatusRequest": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.UserResponse": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/domain.UserRole"
                }
            }
        },
...
        "response.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}